Add logical framework in Challenge2023Day07
The logic for the challenge of day 7 in AdventOfCode is initiated in Kotlin. The approach of determining the value of different characters is set by using a map. The 'solve1' function is now capable of parsing the input lines, determining the current hand and counting the frequency of each number in the hand.