Refactor code to classify poker hands
The main functionality of the code has been refactored to handle multiple poker hands. Individual poker hands are now represented as data class instances of "Hand". Each line of input is processed separately and added to a list of hands, extending the previous mapping to explicitly classify the type of hand. This helps to improve the flexibility and readability of the code.